Update the XDG MIME inplementation to support the draft standards for registering default application (Note: does not use the mime association routines, just defaults)
Lumina-config will now use this backend system for default app registrations, but lumina-open does not use it yet. Also, the defaults tab in lumina-config is still not completely updated yet (resulting in some breakage or unusable buttons).

DO NOT BUILD FROM THIS REVISION - Will be fixed very soon but needed a git snapshot