This is a complete adaptation of lumina-fileinfo.
The most important element is that during the loadDEsktopFile we check the content of the file.
In case the file does not contains the mandatory elements (cfr http://standards.freedesktop.org/desktop-entry-spec/desktop-entry-spec-1.1.html) we reject it and ask the user to manually (via an anoter tool) correct it, or to restart from scratch.