Avoid follow symlinks for size information.
Qt follows symlinks in Unix when size function is used. As a result, if
you have several symlinks that points to big files, the information will
be "wrong". It's better to avoid call size on symlinks and lost several
hundred of bytes instead of get the size of the same file several times.

If needed, lstat systemcall can be used to get symlinks real size.