Massive refactor of Fedora/CentOS RPM spec
The RPM spec for Fedora/CentOS has been massively refactored.

Within the limits of supporting both Fedora and CentOS 7,
the structure has been redone so that 32-bit and 64-bit builds
can be done from the same spec, and that it complies with Fedora's
guidelines on how a package should be structured.

The package is also now at parity with the ports and the Debian
packaging in terms of building and installing Lumina Desktop.

The version number has been bumped to 0.8.8 to reflect that
it will be part of the next release.

The DEPENDENCIES file has also been adjusted to reflect reality.