commeting out  broken  resize function
error was reported by community member. it's built fine on my system, but when building from fresh download it fails.  
with this commented out the project will build.  
I need to figure out what exactly is causing the hangup before re-enabling, but dont want to break the program in the meantime.