Better messageboxes for closing the screenshot app.
Instead of a Yes/No question that requires you to read the text of the
message box and decide what to do, let's give the buttons good names so
we know what the command will do.

Inspired by hitting "no" multiple times when I accidently hit the
shortcut key to take a screenshot. (assuming it was asking "do you want
to save").