Update lumina-checkpass with 2 additional options: 1. "-fd <file descriptor>" pass in a file descriptor (such as 0 for standard input) to read password 2. "-f <file path>" pass in a file path to read the password
Then update lumina-desktop-unified to use the "-f" version with a QTemporaryFile.
I was trying to get it to use the "-fd" version, but the file descriptor I am getting from the QTemporaryFile does not seem to work properly - still need to track this down a bit more later but the -f option works fine for now.